feat(Tabs): simplified API
