+++ src/index.js
@@ -306,30 +306,14 @@ const Styleguide = () => {
                       src: require('./components/Callout/docs.md')
                     },
                     {
-                      path: '/scroller',
-                      title: 'Scroller',
-                      imports: {
-                        Scroller: require('./components/Scroller'),
-                        IconButton: require('./components/IconButton'),
-                        NotificationIcon: require('./components/Icons')
-                          .NotificationIcon
-                      },
-                      src: require('./components/Scroller/docs.md')
-                    },
-                    {
                       path: '/tabs',
                       title: 'Tabs',
                       imports: {
-                        Tabs: require('./components/Tabs'),
-                        tabItems: [
-                          { value: '1', text: 'Tab Number One' },
-                          { value: '2', text: 'Tab Two' },
-                          { value: '3', text: 'Tab Three Yay' }
-                        ],
                         ...require('./components/Format'),
-                        IconButton: require('./components/IconButton'),
-                        NotificationIcon: require('./components/Icons')
-                          .NotificationIcon
+                        Scroller: require('./components/Tabs/Scroller'),
+                        TabButton: require('./components/Tabs/TabButton'),
+                        plainButtonRule: require('./components/Button')
+                          .plainButtonRule
                       },
                       src: require('./components/Tabs/docs.md')
                     }
